+++ lib/src/rules_factory.dart
@@ -1,9 +1,11 @@
 import 'rules/base_rule.dart';
 import 'rules/double_literal_format_rule.dart';
+import 'rules/newline_before_return.dart';
 import 'rules/no_boolean_literal_compare_rule.dart';
 
 const _implementedRules = {
   DoubleLiteralFormatRule(),
+  NewlineBeforeReturnRule(),
   NoBooleanLiteralCompareRule(),
 };